notebooks/fedrag/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fedrag"
version = "1.0.0"
description = ""
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "syft-flwr>=0.2.1",
    "numpy>=2.3.0",
    "faiss-cpu>=1.10.0",
    "tqdm>=4.67",
    "sentence_transformers>=5.0.0",
    "jupyterlab>=4.4.6",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "dimitris"

[tool.flwr.app.components]
serverapp = "fedrag.server_app:app"
clientapp = "fedrag.client_app:app"

[tool.flwr.app.config]
# Possible question answer datasets: PubMedQA, BioASQ, MMLU, MedQA, MedMCQA
# you can pass them as follows: "pubmedqa|bioasq|mmlu|medqa|medmcqa"
# the vertical bar separates the different benchmark datasets
server-qa-datasets = "pubmedqa|bioasq"
# Force to only evaluate the first 10 questions from each benchmark
# dataset, if this value is not set, then evaluate all questions
server-qa-num = 10
# Pass corpus as "StatPearls|Textbooks", the corpus will be assigned
# to each client participating in the federation in a Round-Robin fashion
# The supported corpora are:
# PubMed (23.9M docs), StatPearls (9.3k docs), Textbooks (18 docs), Wikipedia (6.5M docs)
# you can pass them as follows: "pubmed|statpearls|textbooks|wikipedia"
# the vertical bar separates the different corpus
clients-corpus-names = "statpearls|textbooks"
# The value of the Reciprocal Rank Fusion. The value will be used
# to merge multiple documents together at the server-side
# If the value is not set then only the retrieval scores are considered
k-rrf = 60
# The k-nearest-neighbor (most relevant) documents to the given search query
# used by clients to retrieve top-k relevant documents and at the server
# to select the final top-k documents after the merge operation
k-nn = 8
# The LLM used by the server to execute the RAG query.
# We assign the Hugging Face name of the LLM,
# e.g., "HuggingFaceTB/SmolLM2-1.7B-Instruct" or "meta-llama/Llama-3.2-1B-Instruct"
# By default, we use Hugging Face's SmolLM2-1.7B model. For Llama models, please
# ensure that you have accepted the terms and conditions, and have been granted access to the model.
server-llm-hfpath = "HuggingFaceTB/SmolLM2-1.7B-Instruct"
# Whether the server should use a GPU to load the llm and run the inference
server-llm-use-gpu = "false"


[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 2
options.backend.client-resources.num-cpus = 2
options.backend.client-resources.num-gpus = 0
